+++ plugins/openai/src/index.ts
@@ -2,4 +2,5 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 export * as realtime from './realtime/index.js';
-export * as llm from './llm/index.js';
+export * from './models.js';
+export { type LLMOptions, LLM, LLMStream } from './llm.js';
