--- plugins/openai/src/llm/index.ts
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: 2024 LiveKit, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './models.js';
-export { type LLMOptions, LLM, LLMStream } from './llm.js';
